+++ packages/lib/src/lib/map/floors/FloorsHtmlLabels.tsx
@@ -1,0 +1,90 @@
+/* eslint-disable functional/no-expression-statements */
+
+import { Fragment, useRef, type CSSProperties, type ReactNode } from 'react'
+
+import { type LabelText } from '../../../types'
+import { useZoomStyleRef } from '../../viewer/layout/style'
+import type { FloorProps } from './types'
+
+export function RenderFloorLabels({
+  idx: fidx,
+  labels,
+}: FloorProps): ReactNode {
+  const ref = useRef(null)
+  useZoomStyleRef(ref, `labels-${fidx}`)
+  return (
+    <div ref={ref} className="labels">
+      {labels?.map((_text, idx) => (
+        <Fragment key={idx}>
+          <RenderFloorLabel _text={_text} />
+        </Fragment>
+      ))}
+      <style>{labelsStyle}</style>
+    </div>
+  )
+}
+
+type LabelProps = Readonly<{
+  _text: LabelText
+}>
+
+function RenderFloorLabel({ _text }: LabelProps): ReactNode {
+  return (
+    <div
+      className="label"
+      style={
+        {
+          '--x': (_text.attrs?.['x'] || 0) + 'px',
+          '--y': (_text.attrs?.['y'] || 0) + 'px',
+          '--rotate': (_text.attrs?.['rotate'] || 0) + 'deg',
+          '--scale2': (_text.attrs?.['scale2'] || 1) + '',
+          '--scale1': (_text.attrs?.['scale1'] || 1) + '',
+        } as CSSProperties
+      }
+    >
+      {_text.children?.map((_tspan, idx2) => (
+        <p key={idx2}>{_tspan.text ?? ''}</p>
+      ))}
+    </div>
+  )
+}
+
+const labelsStyle = `
+@property --zoom {
+  syntax: '<number>';
+  inherits: false;
+  initial-value: 1;
+}
+div.labels {
+  /* default */
+  --zoom: 1;
+  --zoom-zoom: 1;
+}
+div.label {
+  position: absolute;
+  transform-origin: left top;
+  /*
+  transform: translate(var(--x), var(--y)) rotate(var(--rotate)) scale(var(--zoom)) scale(calc(1 / var(--zoom-zoom))) scale(var(--scale2)) scale(var(--scale1)) translate(-50%, -50%);
+  */
+  transform: translate(var(--x), var(--y)) rotate(var(--rotate)) scale(var(--zoom)) scale(var(--scale2)) scale(var(--scale1)) translate(-50%, -50%);
+  text-align: center;
+  font-family: 'Noto Sans JP', 'Noto Sans', 'sans-serif' !important;
+  font-weight: 200 !important;
+  & > p {
+    margin: 0;
+  }
+}
+/*
+div.labels.zooming > div.label {
+  animation: xxx-label-scale 500ms ease;
+}
+@keyframes xxx-label-scale {
+  from {
+    --zoom: 1;
+  }
+  to {
+    --zoom: var(--zoom-zq-inv);
+  }
+}
+*/
+`
